(feature) Exposing the AG Grid table's getSelectedRows function.
It is very complicated to have to search to find out how to get the data of the rows selected by the user.

For this reason, I decided to implement a method that internally calls the getSelectedRows function of the Ag Grid API, returning the list of data, or, if the value "single" is passed, the first result, and None, if nothing is selected.