QTable: implementation of Quasar's Qtable element
After implementing some elements (my other PRs), I managed to understand the minimum to 'correctly' implement this element.

At the moment, the 'filter/search' API doesn't seem like the best choice, but I don't know how to make it better.

About `qtable.Column`, I didn't know how to define which fields are needed and which fields can be ignored in such a complex object, so I decided to make a wrapper class, which would handle everything.

About the use of `Eval` in Fields: I don't know how dangerous it can be to use eval, however, this was the only method I could think of to transform JS functions 'from string to real objects'.