QTable: fixed: unable to update buffer used by filter
When executing the filter, the current value of `_props['rows']` was saved in `saved_data`, which would later be modified with the filtered values.

However, if during the existence of the QTable element, the values ​​of `_props['rows']` are changed, the filter would not be updated.

Which, firstly, would make it impossible to filter the new values, and secondly, whenever the filter `query_string` becomes empty again, `_props['rows']` would be *updated* to the original value.

Ps.: now, to manipulate the values ​​of the table, you must use:
- .get_data() -> to receive the list of values
- .set_data() -> to update the list of values, including the buffer used in the filter