QTable: activating built-in filter
Now, the search system is Quasar's own, without the need for the Dev to create their own.

There is the `filter_config` object, which accepts some arguments for customizing the search bar.

However, I don't know how to expose which configurations are possible.

I was finding the use of `ui.qtable.Collum` strange, so I decided to go back to the list of objects.

I removed `get_data`, because until then, the Nicegui API does not easily expose a way to get the current data from the table.

However, I kept `set_data`, because whenever I update the table data, it is necessary to run the `evalFunction` function to correctly represent the data.

As I don't know if it's easy for DEV's to remember that, I prefer to expose a simple API.