HTMLTable: implemented pure html table
The creation of this component was requested in #441 issues.

The component is simple:

- columns -> is used to define the thead, it is a list of strings.
- rows    -> is a list of dictionaries, which is expected to have the same amount of keys as the amount of values ​​defined in the columns property.
- style   -> is a dictionary used to enable the definition of arbitrary css for the table.

The sync_config method is used to activate the parsers, which are responsible for transforming lists and dictionaries into html code.

You must always call sync_config (instead of update) when modifying any of the values ​​(columns, rows, style) so that the table is regenerated.