WIP: improve interaction with Matplotlib
This is a sketch of how to improve the interaction with Matplotlib.

The first idea is to directly create `Figure` objects rather than relying on
pyplot.  This de-couples the UI from the global state that is the pyplot figure
registry and ensures that you never get surprise GUI windows popping up (or a
memory leak) when pyplot selects a GUI backend.

The second suggestion is to have the object return by the context manager be
able to create multiple figures (not actually implemented).

The third is to support saving the figures as png instead of svg (not actually
implemented).