update Github Action to cache docker to different locations for each platform (#4051)
Since the docker caching in the GitHub actions still isn't working
properly, this attempts to fix it. As described in [this
PR](https://github.com/zauberzeug/rosys/pull/232) in RoSys the issue is
the multi platform setup in our docker build and push action. This PR
attempts to fix this by using different build caches for each platform.