Prevent `ui.leaflet` from oscillating between two locations (#4121)
This PR tries to solve #4061 by disabling update messages when new
center and zoom values have been received from the client. This
technique is also used for our `ValueElement`s.

A minor downside of this approach might be that maps on the auto-index
page might not stay in sync between different browsers. But I think this
is a reasonable tradeoff.