Add message history and retransmission (#3199)
This PR attempts to resolve #3143 by adding a message history to
`outbox` and providing for the retransmission of missed messages in
order to resynchronize the client's state during a reconnection. If this
cannot be accomplished, a reload of the page is triggered. The goal of
this is to prevent a connected client's state from ever being out of
sync with the server.

For the auto-index page, a history duration of 30 seconds was
arbitrarily chosen. Since this value only determines when the UI is
updated through resending messages instead of a page reload, the UI
should stay properly synchronized regardless of this value.

For a `ui.page`, the history duration is computed based on the expected
lifetime of the `client` object. Currently, with the default
`reconnect_timeout = 3.0`, this is a max of 9 seconds. With this change,
a re-evaluation of this default could be warranted. Now that UI state
can be resynchronized indefinitely, discarding the user's page after
only 5-9s of disconnection seems premature. See
https://github.com/zauberzeug/nicegui/issues/3143#issuecomment-2151376569
for more.

---

Open tasks (October 24, 2024):

- [x] `message_history_length` isn't being used
- [x] handle reconnect when next message ID has already been pruned
- [x] fix failing pytests
- [x] fix test_no_object_duplication_on_index_client
- [x] Should the auto-index client reload when trying to reconnect
(because there is no message history)? -> No.
- [x] ack message to keep history short
- [x] thorough test
- [x] test On Air

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>
Co-authored-by: Rodja Trappe <rodja@zauberzeug.com>