Fix leaked semaphore object warning (#4132)
This pull request fixes #4131: a "leaked semaphore object" warning
appears when aborting with CTRL-C in certain configurations. Semaphores
are used internally in `multiprocessing.Queue` which never get cleaned
up properly.

### Scenario 1: pressing CTRL-C in an FastAPI `app` stared via uvicron
from the command line with auto-reloading

Because the `multiprocessing.Queue` objects in `native.py` have been
created globally they where also instantiated when not in native mode.
This PR fixes this by only creating the objects when native mode is
activated.

### Scenario 2: pressing CTRL-C when using `ui.run(native=True,
reload=True)`

This is still unsolved. I verified that the shutdown/cleanup function
for the Queue objects are correct. But the it seems that the pywebview
window produces also leaked semaphores when pressing CTRL-C.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>