Introduce redis storage which syncs between multiple instances (#4074)
This PR tries to solve
https://github.com/zauberzeug/nicegui/discussions/1606 by implementing a
redis variant of the persistent dictionaries currently used for
`app.storage`. This PR will not solve every remote-syncing-problem out
there. It is meant as a transparent (eg. non-api-breaking) drop in
replacement for the existing local storage mechanism.

NOTE: this is in very early development.

**ToDos**

- [x] make redis server url configureable via `NICEGUI_REDIS_URL`
environment variable, if none is given we use the local storage solution
- [x] think about an elegant way to not sync the whole data set (and
hence decide if we can do this in the scope of this PR) -> we will
implement more sophisticated solutions with mongodb or etcd in other PRs
- [x] make it possible to use `app.storage.user` with redis
- [x] introduce optional redis dependency in `pyproject.toml`
- [x] how do we cope with the neccessary async init? -> by doing the
necessary async init in the page decorator
- [x] fix tests
- [x] make it possible to use `app.storage.tab` with redis
- [x] make db prefix configurable?
- [x] documentation
- [x] fix demo
- [x] mention limitations
- [ ] clean up the example (remove Dockerfile, use newest NiceGUI image
with redis support)

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>