dynamically mirror 'Other HTML elements' demo (#4232)
In PR https://github.com/zauberzeug/nicegui/pull/4027 it was decided to
copy&paste the other HTML section because we want the users to find it
both under "text elements" as well as when looking into the ui.html
element. Duplicating the section isn't a very nice solution though.
Maybe this PR is a better idea for keeping the information in both
places?

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>