Enable usage of leaflet plugins (#2767)
This change enables the usage of leaflet plugins by accepting an
additional argument to leaflet, which is as list of additional
resources (.js and .css) to load after leaflet but before the
map is created.

A small example is included to demonstrate and test the functionality
using the leaflet-rotatedmarker plugin.