Only delete page content if the number of connections is zero (#4285)
This PR tries to solve #4253 with a new `_num_connections` counter. On
handshake the counter is increased, on disconnect it is decreased. On
disconnect, after awaiting the reconnect timeout, the page content is
only deleted if the number of connections is zero.

Possible scenarios:

- normal reconnect:
    ```
handshake (num=1) ... disconnect (num=0) ... handshake (num=1) ... after
delay: don't delete
    ```
- strange reconnect with handshake before disconnect:
    ```
handshake (num=1) ... handshake (num=2) ... disconnect (num=1) ... after
delay: don't delete
    ```
- normal disconnect:
    ```
    handshake (num=1) ... disconnect (num=0) ... after delay: delete
    ```

While working on the handshake and disconnect implementation, I noticed
an inconsistency with `app.on_connect` and `app.on_disconnect`: While
connect handlers are called on reconnect, disconnect handlers are not.
In this PR I changed it so that both handlers are called during
reconnection. We should clarify if this is the desired behavior.