Fix upload to zenodo by using our own built version (#4303)
Unfortunately, there hasn't been any progress on [this
issue](https://github.com/megasanjay/upload-to-zenodo/issues/58), so
this PR tests if the new version that was created by running `npm build`
fixes our problem with the upload.

If this works, we will create another pull request on the original
`upload-to-zenodo` repo.