Add zh-CN & zh-TW support when server down. (#4324)
When the client fails to connect to NiceGUI, a prompt will appear in the
lower left corner of the client: `Connection Lost. Trying to
reconnect...`. This is not difficult to understand for those who know
English. However, for those who do not understand English, it
undoubtedly reduces the user experience.

This pull request adds support for Simplified Chinese and Traditional
Chinese. After using `ui.run(language='zh-CN')` and the browser is fully
loaded, if the server is closed, the correct display on the user's
browser will be `连接丢失` instead of `Connection Lost`. If an undefined
language is set， the default English prompt message will be displayed. I
am still thinking of a better way for developers to customize this
content.

> ui.run(language='zh-CN')

![zh-CN](https://github.com/user-attachments/assets/7c38efa1-4798-465d-b68a-484d5b2f302a)

> ui.run(language='zh-TW')

![zh-TW](https://github.com/user-attachments/assets/12ed2f86-8dc8-493f-89a6-8fe948fd430f)

> ui.run() # Specify another language or do not specify any language

![en-US](https://github.com/user-attachments/assets/0d1034ac-26c1-402d-a960-82196f951979)

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>