Run init of app.storage.general as early as possible (#4355)
When introducing Redis storage in #4074, the initialization of
persistent storage was made async to not block the main thread when
loading data. For that to work the init must be done async -- which
introduced #4352. This PR fixes the issue by moving the init of
`app.storage.general` from app startup to import time and using
`asyncio.run` to create a temporary event loop for the init.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>