Ensure valid host when opening native window (#4443)
This PR fixes #4438, which describes that using `ui.run(host='0.0.0.0',
native=True)` is currently broken. The problem is that `0.0.0.0` is an
placeholder to tell the operating system to bind to all interfaces. But
`0.0.0.0` is not a valid IP to be accessed. So we need to make sure to
not pass it to the native browser window.