let add_css, add_scss and add_sass set shared=True (#4537)
TL-DR: As said in the title. 

---

I am not sure why it wasn't added in the first place. 

`shared=True` was added in this commit:
https://github.com/zauberzeug/nicegui/commit/a09e713ece9099bc2d7164894d2d17d646b1d618,
which was in
[v1.4.6](https://github.com/zauberzeug/nicegui/releases/tag/v1.4.6)

`ui.add_css` was added in this commit:
https://github.com/zauberzeug/nicegui/commit/3e5f92b23c5c7dfa9ab7306c52c84236e5fc0a34
([v1.4.20](https://github.com/zauberzeug/nicegui/releases/tag/v1.4.20)),
with `ui.add_style` (the precursor) in
https://github.com/zauberzeug/nicegui/commit/a0c49a9112cfd04bc8d83d61ea5fc224ab52c95f
([v1.4.19](https://github.com/zauberzeug/nicegui/releases/tag/v1.4.19))

So the function which used `add_head_html` was added after the addition
of `shared=True`? Why not expose all parameters of `add_head_html` since
day1, if `ui.add_css` was just a wrapper around it which opens a
`<style>` tag for you? 🤔 Or am I missing something here...

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>