Pytest: JavaScript calls should not time out when using the user fixture (#4512)
This PR tries to fix #4508 which discovered that a javascript call is
required to create `ui.drawer`. If the test lasted longer than 1 sec the
error handler was called with a javascript timeout. To fix this this PR
does the following:

- implement a general test failure if there are error logs (this makes
the original issue visible in the first place)
- provide a generic mechanism in the user fixture to define what the
simulated response to a certain javascript call should be
- use the mechanism to provide a proper javascript response for a value
check inside `ui.drawer`

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>