Avoid RuntimeError while iterating over bindable properties (#4550)
This PR tries to fix #4419 by disabling the GC while copying the
dictionary of bindable properties. Calling `list()` iterates over the
dictionary, which can case a "RuntimeError: dictionary changed size
during iteration" if the garbage collector runs at the same time and
causes the finalizers to pop items from the dictionary.

The PR is still a draft because there is another iteration over
`bindable_properties` in the `_make_copyable` function. So technically
there is still a chance for the RuntimeError.

1. So we might need to disable the GC once again in `_make_copyable`.
2. Or shouldn't the finalizers pop themselves, but add to a set of
obsolete keys instead?
3. Alternatively we could try to use some kind of lock.