Add theme support to ECharts (#4596)
This PR introduces the ability to specify a custom theme with an ECharts
element.

Users can create a JSON theme file using the ECharts Theme Builder here:
https://echarts.apache.org/en/theme-builder.html
The JSON can then be added to a static files path and the URL passed to
the element or a dictionary (matching the format of the JSON theme) can
be passed directly.

Themes are applied to an instance so multiple charts can be displayed on
the same page with different themes.

Example: 
The first chart has a theme passed as a dict, the second is a URL to a
JSON in static files, the third has no theme specified.

![image](https://github.com/user-attachments/assets/56621e39-c014-477a-876e-38d22c18eab2)

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>