Use a WeakValueDictionary instead of custom finalizers (#4619)
This PR tries to solve the memory leak #4617 as well as the RuntimeError
#4419 using a WeakValueDictionary as proposed by @polyipseity. With the
code from
https://github.com/zauberzeug/nicegui/issues/4617#issuecomment-2804047378
the memory doesn't grow anymore. And the RuntimeError shouldn't be
possible without our custom finalizer.

I think there might still be leaks related to the `bindings` dictionary
which holds strong references to source and target objects. But the
problems described by the two issues mentioned above should be fixed.

@chriswi93 @evnchn @polyipseity @x2a-packet Would you like to have a
look and/or try it out?