Introduce a `_propagation_context` to keep track of visited properties (#4628)
This PR tries to solve #4626 with a new context manager
`_propagation_context` that keeps track of visited properties using a
ContextVar `propagation_visited`. The new implementation better
encapsulates the logic around the `visited` set. This way it simplifies
other parts of the code as they don't have to deal with this set
anymore.

It also fixes the incorrect check
```py
if source_obj_id in visited:
    return
```

@evnchn, @sfeltman Would you like to have a look?