Added documentation on how native app settings are ignored when using a main guard (#4658)
Following discussion in [Pull Request
4627](https://github.com/zauberzeug/nicegui/pull/4627), it was concluded
that we should change documentation instead of code to deal with the
fact that native mode settings are not applied if they come from behind
a main guard.

This PR adds documentation to the [Configuration &
Deployment](https://nicegui.io/documentation/section_configuration_deployment#native_mode)
page, along with an example of bad code and good code.

Feel free to remove the examples if they take up too much space or are
unclear.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>