Fix multi-cursor updates in `ui.codemirror` (#4702)
This PR fixes issue #4701 which has been introduced in PR #4676. When
one change caused the content to shift, subsequent changes have been
applied at the wrong position. This is fixed by creating a new string
from scratch instead of manipulating the old string "in place". While at
it, this PR replaces the many instantiations of large strings with a
single `''.join()` operation to connect individual parts.