run.py: little refactoring to gain basic android compatibility (#4683)
This PR refactors the initialization of the ThreadPoolExecutor and the
ProcessPoolExecutor in `run.py` a little. As a result they get
initialized on the first call to `io_bound` respectively `cpu_bound` on
demand.

This is necessary to achieve basic android compatibility (#578) since
named semaphores are not available on android which the named Executors
use and therefor throw an exception on initialization.

With this change nicegui runs on android under the following conditions:
- calling `run.io_bound` and `run.cpu_bound` will throw exceptions
(named semaphores not available)
- host must not equal '0.0.0.0' for now (because otherwise `welcome.py`
makes an `io_bound` call)

`run.io_bound` and `run.cpu_bound` support might follow in a separate
PR.

With this PR the basic example app at https://github.com/jeff-dh/niceApp
works on android (at least in my environment)

---------

Co-authored-by: Rodja Trappe <rodja@zauberzeug.com>