Sitewide index for AI (#4688)
This PR fix #3412 by serving another 2 JSON files in addition to the
existing `search_index.json`.

- /static/sitewide_index.json: Compared to `search_index.json`, it has
the code of the documentation pages, but it does not have the examples

- /static/examples_index.json: Just the examples as we would otherwise
see in `search_index.json`

This greatly benefits AI as they can learn about the syntax of NiceGUI,
and no need to worry about 2 mixed items in the `search_index.json`
muddying the waters and making AI difficult to learn from the pattern.

---

There should not be any impact on the existing NiceGUI website.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>