Improve On Air reconnect behavior (#4724)
We noticed that an `socketio.AsyncClient` expects you to call
`disconnect` if a connection error occurs. Otherwise `client.connected`
might still be `True` and our re-connect timer early-exits.

Apart from that we noticed that air.py uses a `ui.timer` instead of
`app.timer`. This is ok because it is instantiated on the auto-index
page which lives forever, but it is a smell.