Let npm.py run on Windows as well (hopefully same function as old) (#4733)
This PR lets npm.py run on Windows as well. It also allows it to be ran
repeatedly without throwing errors

Core changes: 
- Use `tempfile.TemporaryDirectory()`
- `unlink` before `rename` so that the old file is overwritten
- Try-except the failing line `if not any(destination.iterdir()):`
`destination.rmdir()`

---

Admittedly, this is a minor concern, because by all means you should be
on Linux or using a Dev Container, according to the [contributing
guide](https://github.com/zauberzeug/nicegui/blob/42d24d40129785038d6fce9c2aafe24c62482478/CONTRIBUTING.md).

But, both has failed me so far, and so I'm still on Windows. And, I need
`npm.py` to work, so as to help out in #4656 and #4520.

I personally believe that making `npm.py` runnable on cross-platform may
speed up development of new elements, as we lower the barrier to custom
element development.

---

It is best to answer the following before we proceed: 

- [ ] What's the purpose of the `if not any(destination.iterdir()):`
`destination.rmdir()`?
  - It seems to always fail on Windows, so I add a try-except
  - But I don't want to just carelessly do that and call it a day
- [x] Is the output on Linux still working? 
- You may want to grab the `npm.json` with pinned version in #4632, and
test.
- It's the same on my side, minus Tailwind, which they changed stuff
upstream and so we'll never get the exact same copy.