Get ruff happy after tightning the single-quote rules (#4788)
### Motivation

This PR fix #4787 by changing our code around how Ruff and flake8 works,
such that it does not complain on GitHub Actions.

### Implementation

The code is made simply by: 
- `pre-commit run --all-files`
- Commiting whatever was changed

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).

### Special notes

As noted in
https://github.com/zauberzeug/nicegui/issues/4787#issuecomment-2905344451,
I personally believe only 1 out of 3 siutations, that it is our firm
mistake. Other cases, I think its not so much. I think "docstring-like
comments made with double-quote strings" is fine for clarity, while
"docstring made with single-quote strings" is negotiable.

But, without merging this PR or some other substitute to get Ruff happy,
contributors's action runs will always fail, which is **a massive
pain**, so I would prefer that we maybe merge this before we discuss
whether we should and how to loosen Ruff rules a bit