Introduce "Dynamic Resources" for robust inclusion of `codehilite.css` against match-all `@ui.page("/{_:path}")` (#4778)
This PR fix #4774 by introducing the concept of "Dynamic Resources",
resources whose `Response` (and thus content) are provided as return
value by a callable.

This is a new and unique mechanism compared to
libraries/components/resources, which are all static and
served-from-the-file.

#### Reason for dynamic generation

This is because we have established that we'd like to keep dynamically
generating the Codehilite CSS such that pygments styling work
appropriately in
https://github.com/zauberzeug/nicegui/issues/4774#issuecomment-2897824724

#### Code implementation highlights

This implementation is much better than the
`core.app.get(CODEHILITE_CSS_URL)(lambda: PlainTextResponse(...)`, in a
sense that:

- Properly uses `app.get` as a decorator, instead of a sketchy way. 
- Definition and serving of dynamic resources is centralized. 
- Works against match-all `@ui.page("/{_:path}")`, as much as the rest
of the resources work.

The dynamic resources are available at
`f'/_nicegui/{__version__}/dynamic_resources/{{filename}}'`. Note that
there is no `key`, since we don't need to do cachebusting since we
should always serve these resources with caching turned off.

#### Notable point

In #4540, a middleware will be introduced. We need to modify it so that
`dynamic_resources` does not get the long `cache_control` header.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>